refactor: merge internal/server into internal/config
- Remove internal/server package
- Consolidate logic into internal/config
