+++ internal/app/handlers_test.go
@@ -6,8 +6,8 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"watchdog.onebusaway.org/internal/config"
 	"watchdog.onebusaway.org/internal/models"
-	"watchdog.onebusaway.org/internal/server"
 )
 
 func TestHealthcheckHandler(t *testing.T) {
@@ -57,7 +57,7 @@ func TestHealthcheckHandler(t *testing.T) {
 	})
 
 	t.Run("returns 500 when no servers configured", func(t *testing.T) {
-		cfg := server.NewConfig(
+		cfg := config.NewConfig(
 			4000,
 			"testing",
 			[]models.ObaServer{},
